60 ograniczenie ruchu jednostek from main

+++ desktop/src/com/mygdx/game/client/ecs/system/MovementSystem.java
@@ -5,10 +5,13 @@ import com.artemis.annotations.All;
 import com.artemis.systems.IteratingSystem;
 import com.mygdx.game.client.ecs.component.Highlighted;
 import com.mygdx.game.client.model.ChosenEntity;
+import com.mygdx.game.client.ui.OutOfMoveRangeDialogFactory;
 import com.mygdx.game.client_core.ecs.component.Movable;
 import com.mygdx.game.client_core.network.MoveEntityService;
 import com.mygdx.game.core.ecs.component.Coordinates;
 import com.mygdx.game.core.ecs.component.Field;
+import com.mygdx.game.core.ecs.component.Stats;
+import com.mygdx.game.core.util.DistanceUtil;
 import lombok.NonNull;
 import lombok.extern.java.Log;
 
@@ -20,26 +23,43 @@ public class MovementSystem extends IteratingSystem {
 
   private final ChosenEntity chosenEntity;
   private final MoveEntityService moveEntityService;
+  private final OutOfMoveRangeDialogFactory moveRangeDialog;
   private ComponentMapper<Highlighted> highlightedMapper;
   private ComponentMapper<Coordinates> coordinatesMapper;
+  private ComponentMapper<Stats> statsMapper;
   private ComponentMapper<Field> fieldMapper;
 
   @Inject
   public MovementSystem(
       @NonNull ChosenEntity chosenEntity,
-      @NonNull MoveEntityService moveEntityService
+      @NonNull MoveEntityService moveEntityService,
+      @NonNull OutOfMoveRangeDialogFactory moveRangeDialog
   ) {
     this.chosenEntity = chosenEntity;
     this.moveEntityService = moveEntityService;
+    this.moveRangeDialog = moveRangeDialog;
   }
 
   @Override
   protected void process(int entityId) {
     if (chosenEntity.isAnyChosen() && fieldMapper.has(chosenEntity.peek())) {
       log.info("some are chosen and there's a movable highlighted entity");
+
       var targetCoordinate = coordinatesMapper.get(chosenEntity.pop());
-      moveEntityService.moveEntity(entityId, targetCoordinate);
-      highlightedMapper.remove(entityId);
+      var currentCoordinate = coordinatesMapper.get(entityId);
+
+      var distance = DistanceUtil.distance(currentCoordinate, targetCoordinate);
+      var range = statsMapper.get(entityId).getMoveRange();
+
+      log.info("Chce przejsc " + distance + " ale zostało mi " + range);
+
+      if (distance > range){
+        moveRangeDialog.createAndShow(range, distance);
+      }
+      else {
+        moveEntityService.moveEntity(entityId, targetCoordinate);
+        highlightedMapper.remove(entityId);
+      }
     }
   }
